Complete profile when one command is in progress
Currently profile completes without arguments, but as soon as you have
written p and try to complete the completion will disappear and have no
match. This fixes this by also including it when a command is
provided and fish will filter it out if it does not match the current
token.