Complete in progress commands
Pass the content up to the previous token so you get completions also
for the token in progress. Without this the completion disappears when
you reach a single match. For example `aws ssm get-parameter` won't
match `aws ssm get-parameters`. There is one caveat that
`aws_completer` won't return option completions for a command unless you
start with `-`. This passes a `-` to the previous token in case the last
token starts with a `-`.